+++ src/main/java/lotto/LottoMachine.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.lotto.LottoNumbers;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
