+++ src/main/java/lotto/lotto/Condition.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.lotto;
 
 @FunctionalInterface
 public interface Condition {
