+++ src/main/java/lotto/lotto/Lotto.java
@@ -1,28 +1,22 @@
-package lotto;
+package lotto.lotto;
 
 import java.util.List;
 import java.util.Objects;
 
 public class Lotto {
     private final LottoNumbers lottoNumbers;
-    private final long matchedLottoNumbersCount;
 
-    private Lotto(LottoNumbers lottoNumbers, long matchedLottoNumbersCount) {
+    private Lotto(LottoNumbers lottoNumbers) {
         this.lottoNumbers = lottoNumbers;
-        this.matchedLottoNumbersCount = matchedLottoNumbersCount;
     }
 
     public static Lotto from(List<Integer> input) {
         LottoNumbers lottoNumbers = LottoNumbers.from(input);
-        return new Lotto(lottoNumbers, 0);
+        return new Lotto(lottoNumbers);
     }
 
     public static Lotto from(LottoNumbers input) {
-        return new Lotto(input, 0);
-    }
-
-    public static Lotto from(LottoNumbers input, long matchedLottoNumbersCount) {
-        return new Lotto(input, matchedLottoNumbersCount);
+        return new Lotto(input);
     }
 
     public List<Integer> lottoNumbers() {
