+++ src/main/java/lotto/result/LottoResults.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.result;
 
 import java.util.Map;
+import java.util.Objects;
 
 public class LottoResults {
     private final Map<MatchedNumbers, Long> values;
@@ -8,7 +9,7 @@ public class LottoResults {
 
     public LottoResults(Map<MatchedNumbers, Long> lottoResults) {
         this.values = lottoResults;
-        this.prize = sumPrize(lottoResults);
+        this.prize = totalPrize(lottoResults);
     }
 
     public void add(MatchedNumbers matchedNumbers, long lottosCount) {
@@ -23,7 +24,7 @@ public class LottoResults {
         return prize;
     }
 
-    private float sumPrize(Map<MatchedNumbers, Long> lottoResults) {
+    private float totalPrize(Map<MatchedNumbers, Long> lottoResults) {
         float prize = 0f;
 
         for (MatchedNumbers matchedNumbers : lottoResults.keySet()) {
@@ -31,5 +32,18 @@ public class LottoResults {
         }
 
         return prize;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        LottoResults that = (LottoResults) o;
+        return Float.compare(that.prize, prize) == 0 && Objects.equals(values, that.values);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(values, prize);
     }
 }
