+++ src/test/java/lotto/LottoMachineTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.lotto.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
