+++ src/test/java/lotto/LottosTest.java
@@ -1,5 +1,9 @@
 package lotto;
 
+import lotto.lotto.Lotto;
+import lotto.lotto.LottoNumbers;
+import lotto.lotto.Lottos;
+import lotto.result.MatchedNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
